255BDiscord/requirements.txt
discord.py
python-dotenv
pytz
random